refactor backtracking
