+++ Brute_force/input.txt
@@ -1,19 +1,15 @@
-8 4 3
-1 4 1 2 2 1
-0 9 1 8 7 4 8 9
-2 0 4 6 3 1 3 8
-7 10 0 3 4 6 10 10
-1 1 3 0 6 5 3 1
-7 2 6 3 0 2 9 6
-3 9 5 4 3 0 2 8
-10 7 6 3 10 5 0 6
-6 4 3 2 4 1 1 0
-3 6 10 3
-3 2 2 9
-6 9 4 8
-4 10 2 10
-4 1 8 3
-5 10 7 9
-8 2 9 3
-1 6 1 8
-1 3 4 2 4 4 2 4
+6 4 2
+2 4 1 3 1 1
+0 2 4 1 2 5 
+2 0 5 5 3 5 
+4 5 0 4 3 5
+1 5 4 0 3 2
+2 3 3 3 0 3 
+5 5 5 2 3 0 
+3 5 1 5 
+5 2 5 3 
+3 1 3 3 
+5 5 1 3 
+4 5 4 1 
+5 3 4 5 
+1 3 4 2 2 3
